refactor: 챌린지 페이지 컴포넌트 및 utils 분리
